refactor: 프로필 편집 mutationOptions로 변경
