+++ apps/client/src/shared/constants/mutation-key.ts
@@ -1,0 +1,3 @@
+export const USER_MUTATION_KEY = {
+  EDIT_PROFILE: () => ['edit-profile'],
+} as const;
